Made the Tile Render Engine
The editor can now render the map. This is still a VERY basic system as
it simply renders it in one and the same location all the time, and
doesn't care whether or not some tiles are actually visible on-screen or
not.

Sadly, you can't see the results of this change unless you manually code
in some tiles onto the map. I haven't figured out input just yet.